requirements.txt
flask
pandas
numpy
python-dotenv
google-generativeai